+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## master (unreleased)
 
+- Fix problem with running migrations for `activerecord` 8.0.0.rc2
+
 ## 0.20.0 (2024-10-21)
 
 - Enhance columns removal check to check indexes `where` and `include` options